Load available options when no value is set and fix inlines prefixes
When the widget is rendered we show empty_label even if we have no value
available.
In JS when a select is changed, we check if there are 'new' shown
related selects and replace the prefixes correctly (when adding an
inline).
Also we fire a change event on document ready so selects with no value
get available options from their related select.